# packages/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fakegaming Bot API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Login via /auth/login and paste the token here using the \"Authorize\" button."
      }
    },
    "schemas": {
      "BirthdayConfig": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "day": {
            "type": "integer",
            "format": "int64"
          },
          "month": {
            "type": "integer",
            "format": "int64"
          },
          "year": {
            "type": "integer",
            "format": "int64"
          },
          "guildId": {
            "type": "string"
          },
          "channelId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "userId",
          "createdAt",
          "updatedAt"
        ]
      },
      "DisabledCommandConfig": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "guildId": {
            "type": "string"
          },
          "commandName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Login and get a JWT token",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "code": "your_discord_oauth_code"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT token and user info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "username": {
                          "type": "string"
                        },
                        "discriminator": {
                          "type": "string"
                        },
                        "avatar": {
                          "type": "string"
                        }
                      }
                    },
                    "guilds": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "icon": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/birthdays": {
      "get": {
        "summary": "List all birthdays",
        "tags": [
          "Birthdays"
        ],
        "responses": {
          "200": {
            "description": "List of birthdays",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BirthdayConfig"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add or update a birthday",
        "tags": [
          "Birthdays"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BirthdayConfig"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/birthdays/{userId}/{guildId}": {
      "get": {
        "summary": "Get a birthday by userId and guildId",
        "tags": [
          "Birthdays"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "guildId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Birthday config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BirthdayConfig"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Remove a birthday by userId and guildId",
        "tags": [
          "Birthdays"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "guildId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/disabledCommands": {
      "get": {
        "summary": "List all disabled commands",
        "tags": [
          "DisabledCommands"
        ],
        "responses": {
          "200": {
            "description": "List of disabled commands",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisabledCommandConfig"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a disabled command",
        "tags": [
          "DisabledCommands"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisabledCommandConfig"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/disabledCommands/check": {
      "get": {
        "summary": "Check if a command is disabled in a guild",
        "tags": [
          "DisabledCommands"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "guildId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "commandName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Disabled status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "disabled": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/disabledCommands/{id}": {
      "delete": {
        "summary": "Remove a disabled command by id",
        "tags": [
          "DisabledCommands"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/disabledCommands/guild/{guildId}": {
      "get": {
        "summary": "List all disabled commands for a specific guild",
        "tags": [
          "DisabledCommands"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guildId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of disabled commands for the guild",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisabledCommandConfig"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}